+++ packages/shared/src/index.ts
@@ -5,6 +5,7 @@ export * from "./auth";
 export * from "./eden";
 export * from "./connections";
 export * from "./ddl";
+export * from "./engine";
 export * from "./export";
 export * from "./meta";
 export * from "./mutation";